teach: refactor and fix several bugs
